Define images for pages. Including defaults.
A template-tag is used to get the image-urls into the template. A setting is used to define the types of images you want to connect to pages. Defaults for the types can be given in the admin-backend. When no image of the given type is found for a page, the pages parent is search recursively for an appropriate image.

Tests and documentation should come next.